==> 1-GVA_data_scrape_listing_links.ipynb <==
import src.login_config as config
import src.data_scrape as ds

browser = ds.selenium_start()
ds.first_time_login(browser)
lnks = ds.fn_to_retreive_sold_listings(browser, 'richmond',84)

import pandas as pd
table = pd.DataFrame(lnks)

table = table.rename(columns= {0:"Listing link"})
print(len(table))
table = table.drop_duplicates()

table.to_csv("richmond_sold_listings_new.csv")
