==> 2-GVA_retrieve_data_attributes.ipynb <==
import src.login_config as config
import src.helpers as helpers
import src.data_scrape as ds
import pandas as pd

# Retrieve of all sold listing links that was previously scraped in 1-GVA_data_scrape_listing_links notebook
combined_df = helpers.read_csv_files('sold_listings')

print(len(combined_df))

richmond_df = pd.read_csv('richmond_sold_listings_new.csv',index_col = [0])

The code will take some time to run. Therefore, I have split the code to run in batches (by cities). 

# sample. split them by batch (city). With the exception of Vancouver, I have split it into 2 batches due to the immense size.
# To replicate this, rename the dataframe and change the name of the "City. For example, if I want Burnaby then change dataframe to the folllowing:
# burnaby_df = combined_df[combined_df['City'] == 'burnaby'].reset_index(drop = True) "

vancouver_df = combined_df[combined_df['City'] == 'vancouver'].reset_index(drop = True)

van_df1 = vancouver_df[0:2700]
van_df2 = vancouver_df[2700:].reset_index(drop = True)

import time
import pandas as pd

batch = richmond_df

browser = ds.selenium_start()
ds.first_time_login(browser)

df_list = list()
for i in range(0,len(batch)):
    start_time = time.time()
    
    table = ds.retrieve_sold_listing_description(browser, batch['Listing link'][i])
    df_list.append(table)
    print(i)
    end_time = time.time()
    
    execution_time = end_time - start_time
    print("Execution time:", execution_time, "seconds")

full_table = pd.concat(df_list, axis = 0, ignore_index = True)

# Save progress in csv file
# For other cities they are saved in this format: burnaby_complete, coquitlam_complete, etc..)
full_table.to_csv('richmond2_complete.csv')

# Read in all the csv files that was previously run in the code above for all cities
cities_df = helpers.read_csv_files('_complete')

# some listings have none values. let's remove it
cities_df = cities_df[~pd.isnull(cities_df['list_date'])].reset_index(drop = True)
cities_df['City'] = cities_df['City'].replace({'vancouver1': 'vancouver', 'vancouver2':'vancouver'})

#Store this in a database

import pandas as pd
## connect to database
import sqlite3

conn = sqlite3.connect("housing_information.db")

##push the dataframe to sql 
cities_df.to_sql("df_listings", conn, if_exists="replace", index = False)

##create the table

conn.execute(
    """
    create table my_table as 
    select * from df_listings
    """)

# Reading data in from database
pd.read_sql_query('SELECT * FROM df_listings', conn)
