==> 3-Data_enrichment.ipynb <==
import pandas as pd
import sqlite3
import src.data_enrichment as enrich
import src.helpers as helpers
import src.login_config as login_config

# Read in data
conn = sqlite3.connect("housing_information.db")
data_df = pd.read_sql_query('SELECT * FROM df_listings', conn)

## Data enrichment

data_df['house_age'] = data_df['approx_age'].apply(enrich.age_of_house)
data_df['difference_in_days'] = data_df.apply(lambda row: enrich.calculate_days_to_sell(row['list_date'], row['end_date']), axis=1)
data_df['price_difference_abs'] = data_df.apply(lambda row: enrich.calculate_price_diff_or_pct(row['list_price'], row['sold_price'], 'difference'), axis=1)
data_df['price_difference_pct'] = data_df.apply(lambda row: enrich.calculate_price_diff_or_pct(row['list_price'], row['sold_price'], 'percentage'), axis=1)

## Data Exploration

### Address 

Currently the address name is retrieved in the following format:
<br> - "Property Address: {name of address}, {city}, {province}"
<br> - Example: 'Property Address: Juneau Street, Burnaby, British Columbia'
<br>
<br> We only need the {name of address}. We will perform regex in order to retrieve it.

# Clean up "property_addr" field
data_df['Address_Clean'] = data_df['property_addr'].apply(enrich.regex_cleanup)

The following addresses need to be updated. For some odd reason, the data scrape algorithm did not catch the full address of these names. We will need to manually check and verify the correct name of the address.

# few rows that needs to be replaced
# Juneau Street -> 1301 - 4488 Juneau Street
# Soball Street -> 38 - 1295 Soball Street
# Seaham Crescent -> 11300 Seaham Crescent
# Yale Street -> 2558 Yale Street [ But remove this since lot was bought]

data_df.loc[data_df['Address_Clean'] == 'Juneau Street','Address_Clean'] ='1301 - 4488 Juneau Street'
data_df.loc[data_df['Address_Clean'] == 'Soball Street','Address_Clean'] ='38 - 1295 Soball Street'
data_df.loc[data_df['Address_Clean'] == 'Seaham Crescent','Address_Clean'] ='11300 Seaham Crescent'
data_df = data_df[data_df.Address_Clean != 'Yale Street']

### Home type

data_df.house_type.value_counts()

Notice above that house type are categorized into multiple categories. However, the ones with low value counts actually below in one of the following categories:
<br> - Apartment/Condominium
<br> - Single Family Residence
<br> - Townhouse

data_df['Home_type'] = data_df['house_type'] 

data_df.loc[data_df['house_type'] == 'Quadruplex','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Manufactured On Land','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Condo Apt','Home_type'] ='Apartment/Condominium'
data_df.loc[data_df['house_type'] == 'Triplex','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Duplex','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Manufactured Home','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Multi Family','Home_type'] ='House'
data_df.loc[data_df['house_type'] == 'Single Family Residence','Home_type'] ='House'

Let's verify that the changes have indeed been made

data_df.Home_type.value_counts()

### Maintanience fee
Remove it as they are all blanks. We will revisit this later
<br> For Apartment/Condominum, we could fetch this data off of condos.ca. There may be a better way to retrieve maintenacne fee. In the meantime, let's not include this column in the master copy.

data_df.maint_fee.value_counts()

### Levels

data_df.levels.value_counts()

Based on the distribution of the field "levels". It seems that this variable is not that useful as majority of the listings are classified as "Residential". Therefore we will not include this variable in the final dataframe to work with.

### Garage 

data_df.garage.value_counts()

Notice there is one row that lists as "Undergrnd". Will update from "Undergrnd" to "Yes".

data_df.loc[data_df['garage'] == 'Undergrnd','garage'] ='Yes'

Verification

data_df.garage.value_counts()

### House age
There are 6 rows (listings) where house age is not calculated. This it because no built age was provided.

print('The total # of rows before dropping nas is:', len(data_df)) # Before dropping nas
data_df = data_df.dropna(axis = 0, subset = ['house_age']).reset_index(drop = True)
print('The total # of rows after dropping nas is:', len(data_df)) # After dropping nas --  Before dropping nas - After dropping nas = 6 rows

Create master dataframe

columns_to_keep = ['mls_number', 'Address_Clean', 'City', 'Home_type', 'approx_age', 'house_age', 
                   'garage','garage_size','taxes', 'avg_price_sqft', 'bedroom','bathroom','list_date',
                   'list_price','end_date','sold_price', 'difference_in_days','price_difference_abs', 
                   'price_difference_pct']

df_master = data_df[columns_to_keep].copy()

df_master.rename(columns = {'Address_Clean': 'address',
                            'approx_age': 'yr_built',
                            'house_age': 'home_age'}, inplace = True)
df_master.columns = map(str.lower, df_master.columns)

## Features

import googlemaps
gmaps = googlemaps.Client(key = login_config.GEOCODE_KEY)

df_master['geocode_json'] = df_master.apply(lambda row: enrich.geocode_json(row['address'], row['city'], gmaps), axis=1)
df_master['geocode_json'] = df_master['geocode_json'].astype(str)

# #Store this in a database, # saving progress

# # conn = sqlite3.connect("housing_information.db")

# ##push the dataframe to sql 
# df_master.to_sql("df_master", conn, if_exists="replace", index = False)

# ##create the table

# conn.execute(
#     """
#     create table my_table_master as 
#     select * from df_master
#     """)

# # Read data back in
# # Read in data
# conn = sqlite3.connect("housing_information.db")
# df_master = pd.read_sql_query('SELECT * FROM df_master', conn)

# import ast
df_master['geocode_json'] = df_master['geocode_json'].apply(ast.literal_eval)
df_master['postal_code'] = df_master['geocode_json'].apply(enrich.extract_geocode_postal)

df_master['index_col'] = range(0,len(df_master))

There are some rows where the postal code is not length 7. By definition, in Canada, postal code should be in the following format: 
<br> - A1A 1A1, where A is letter and 1 is a digit, with a space separating the third and fourth characters. Therefore the length of the postal code is 7.

filtered_df_equal_7  = df_master[df_master['postal_code'].str.len() == 7]
print(f"The number of rows where postal code == 7 is: {len(filtered_df_equal_7 )}")

filtered_df_not_equal_7 = df_master[df_master['postal_code'].str.len() != 7]
print(f"The number of rows where postal code != 7 is: {len(filtered_df_not_equal_7 )}")

There are a total of 1077 rows (listings) where the postal codes do not match the format of a canadian postal code. Let's explore.

Example #1:

filtered_df_not_equal_7['geocode_json'][38] # Doesn't seem like searching address with just street + city is enough.

# Example 2:
filtered_df_not_equal_7['geocode_json'][14911] # seems like there is a similar location in the USA. There is a city in USA called Vancouver

Let's re-run the addresses in filtered_df_not_equal_7, where postal codes did not have length == 7. Let's tweak our function enrich.geocode_json to include provincial and country during the search of the geocode json

filtered_df_not_equal_7['postal_code_correction'] = filtered_df_not_equal_7.apply(lambda row: enrich.geocode_json(row['address'], row['city'], gmaps), axis=1)
filtered_df_not_equal_7['postal_code_correction'] = filtered_df_not_equal_7['geocode_json_rerun'].apply(enrich.extract_geocode_postal)

# rerun_filtered_df_equal_7  = filtered_df_not_equal_7[filtered_df_not_equal_7['postal_code_correction'].str.len() == 7]
# print(f"The number of rows where postal code == 7 is: {len(rerun_filtered_df_equal_7 )}")

# rerun_filtered_df_not_equal_7 = filtered_df_not_equal_7[filtered_df_not_equal_7['postal_code_correction'].str.len() != 7]
# print(f"The number of rows where postal code != 7 is: {len(rerun_filtered_df_not_equal_7 )}")

# rerun_filtered_df_not_equal_3 = rerun_filtered_df_not_equal_7[rerun_filtered_df_not_equal_7['postal_code_correction'].str.len() != 3]
# print(f"The number of rows where postal code != 3 is: {len(rerun_filtered_df_not_equal_3 )}")

df1 = filtered_df_equal_7
cols_to_add = ['mls_number', 'address', 'city', 'home_type', 'yr_built', 'home_age',
       'garage', 'garage_size', 'taxes', 'avg_price_sqft', 'bedroom',
       'bathroom', 'list_date', 'list_price', 'end_date', 'sold_price',
       'difference_in_days', 'price_difference_abs', 'price_difference_pct',
       'geocode_json_rerun', 'postal_code_correction', 'index_col']
df2 = filtered_df_not_equal_7[cols_to_add]
df2 = df2.rename(columns = {'geocode_json_rerun': 'geocode_json',
                  'postal_code_correction': 'postal_code'})
# geocode_json_rerun, postal_code_correction
remaster_df = pd.concat([df1, df2])
remaster_df = remaster_df.sort_values(by = ['index_col'], ascending = True)

remaster_df['geocode_json'] = remaster_df['geocode_json'].astype(str)

# #Store this in a database, # saving progress

# conn = sqlite3.connect("housing_info.db")

# ##push the dataframe to sql 
# df_master.to_sql("df_master", conn, if_exists="replace", index = False)

# ##create the table

# conn.execute(
#     """
#     create table my_table as 
#     select * from df_master
#     """)

import ast
conn = sqlite3.connect("housing_info.db")
df_master = pd.read_sql_query('SELECT * FROM df_master', conn)
df_master['geocode_json'] = df_master['geocode_json'].apply(ast.literal_eval)

# conn = sqlite3.connect("housing_dataframes.db")
# cursor = conn.cursor()
# cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
# print(cursor.fetchall())

# cursor.execute("DROP TABLE my_table_master;")

filtered_df_equal_7  = df_master[df_master['postal_code'].str.len() == 7]
print(f"The number of rows where postal code == 7 is: {len(filtered_df_equal_7 )}")

filtered_df_not_equal_7 = df_master[df_master['postal_code'].str.len() != 7]
print(f"The number of rows where postal code != 7 is: {len(filtered_df_not_equal_7 )}")

Of the postal codes where the length is not equal to 7 (Canadian format). Let's determine what the formats are like

filtered_df_not_equal_7.postal_code

### Postal Code format description
We can see that some postal codes have the first three characters of the postal code. According to the definition of a canadian postal code, the first three characters of a postal determine the following: 
<br> - First letter denotes a particular "postal district", which outside of Quebec and Ontario, corresponds to an entire province or territory.
<br> - Second letter denotes a specific rurual region, an entire medium-sized city or a section of a major metropolitian area.
<br> - Third letter denotes an extract area if it is a city/town or other geographic area
<br> Local Delivery Unit (LDU) makes up the remaining characters of the postal code. This is for determining the specific building, city block for delivery. Therefore, the first 3 characters is sufficient in determining the approximate location.
<br>
<br>For a better depiction of the list of codes, please visit https://www.canadapost-postescanada.ca/cpc/en/support/articles/addressing-guidelines/postal-codes.page#:~:text=Two%20segments%20make%20up%20a,numeric%2Dalpha%2Dnumeric).

Source: https://en.wikipedia.org/wiki/Postal_codes_in_Canada

filtered_df_not_equal_7_3 = filtered_df_not_equal_7[filtered_df_not_equal_7['postal_code'].str.len() != 3]
print(len(filtered_df_not_equal_7_3))

There are 108 rows (or listings) that do not have consist 3 characters in their postal code.

filtered_df_not_equal_7_3[60:80]
# 1806 - 2008 Fullerton Avenue -> V7P

Let's attempt to replace those "None" values to actual postal codes.

import geopy
geo_locator = geopy.Nominatim(user_agent='my_request')
                        # Latitude, Longitude
lat = filtered_df_not_equal_7_3['geocode_json'][14485][0]['geometry']['location']['lat']
lon = filtered_df_not_equal_7_3['geocode_json'][14485][0]['geometry']['location']['lng']
r = geo_locator.reverse((lat, lon))
print(r.raw['address']['postcode'])

print(lat)
print(lon)

For some reason, this seems to be pointing to a specific building and multiple addresses are doing. This is not accurate. Since there are only 108 rows, let's try to manually fill the information.

enrich.regex_fix_misc(df_master)

df_master.loc[df_master['mls_number'] == 'R2818744','postal_code']

r

# geocode_result = gmaps.geocode(df_master.address[0])

import importlib
importlib.reload(enrich)
